+++ packages/a11y/package.json
@@ -1,0 +1,15 @@
+{
+	"name": "@ag.ds-next/a11y",
+	"version": "1.0.0",
+	"main": "dist/ag.ds-next-a11y.cjs.js",
+	"module": "dist/ag.ds-next-a11y.esm.js",
+	"license": "MIT",
+	"peerDependencies": {
+		"@emotion/react": "^11.7.0",
+		"react": "^16.14.0 || ^17.0.0"
+	},
+	"devDependencies": {
+		"@emotion/react": "^11.7.0",
+		"react": "17.0.2"
+	}
+}
