+++ packages/search-box/package.json
@@ -8,6 +8,7 @@
 		"@reach/auto-id": "^0.16.0"
 	},
 	"peerDependencies": {
+		"@ag.ds-next/a11y": "1.0.0",
 		"@ag.ds-next/box": "3.0.0",
 		"@ag.ds-next/button": "3.0.0",
 		"@ag.ds-next/core": "2.0.0",
@@ -18,6 +19,7 @@
 		"react": "^16.14.0 || ^17.0.0"
 	},
 	"devDependencies": {
+		"@ag.ds-next/a11y": "*",
 		"@ag.ds-next/box": "*",
 		"@ag.ds-next/button": "*",
 		"@ag.ds-next/core": "*",
