+++ packages/side-nav/package.json
@@ -5,12 +5,14 @@
 	"module": "dist/ag.ds-next-side-nav.esm.js",
 	"license": "MIT",
 	"peerDependencies": {
+		"@ag.ds-next/a11y": "1.0.0",
 		"@ag.ds-next/box": "3.0.0",
 		"@ag.ds-next/core": "2.0.0",
 		"@emotion/react": "^11.7.0",
 		"react": "^16.14.0 || ^17.0.0"
 	},
 	"devDependencies": {
+		"@ag.ds-next/a11y": "*",
 		"@ag.ds-next/box": "*",
 		"@ag.ds-next/core": "*",
 		"@emotion/react": "^11.7.0",
