+++ packages/side-nav/src/SideNavLink.tsx
@@ -2,9 +2,10 @@ import { ReactNode } from 'react';
 import { Box } from '@ag.ds-next/box';
 import { useLinkComponent, mapSpacing } from '@ag.ds-next/core';
 import { boxPalette, fontGrid, outline } from '@ag.ds-next/core';
+import { VisuallyHidden } from '@ag.ds-next/a11y';
 
 import { useLinkListDepth } from './context';
-import { localPalette, visuallyHiddenStyles } from './utils';
+import { localPalette } from './utils';
 
 export type SideNavLinkType = {
 	active?: boolean;
@@ -81,7 +82,7 @@ export const SideNavLink = ({
 				<Link href={href}>
 					{label}
 					{isCurrentPage ? (
-						<span css={visuallyHiddenStyles}> Current page</span>
+						<VisuallyHidden> Current page</VisuallyHidden>
 					) : null}
 				</Link>
 			</Box>
