+++ packages/side-nav/src/SideNavTitle.tsx
@@ -1,8 +1,7 @@
 import { ReactNode } from 'react';
 import { Box } from '@ag.ds-next/box';
 import { boxPalette, useLinkComponent } from '@ag.ds-next/core';
-
-import { visuallyHiddenStyles } from './utils';
+import { VisuallyHidden } from '@ag.ds-next/a11y';
 
 export const SideNavTitle = ({
 	isCurrentPage,
@@ -35,9 +34,7 @@ export const SideNavTitle = ({
 			}}
 		>
 			{children}
-			{isCurrentPage ? (
-				<span css={visuallyHiddenStyles}> Current page</span>
-			) : null}
+			{isCurrentPage ? <VisuallyHidden> Current page</VisuallyHidden> : null}
 		</Box>
 	);
 };
